# docs/swagger.yaml
definitions:
  model.Article:
    properties:
      content:
        type: string
      id:
        type: integer
      type:
        type: string
    type: object
  model.Result:
    properties:
      code:
        example: 0
        type: integer
      data:
        type: object
      message:
        example: 请求信息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: youngxhui@gmail.com
    name: youngxhu
    url: https://youngxhui.top
  description: Gin swagger 示例项目
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin swagger
  version: "1.0"
paths:
  /article:
    post:
      consumes:
      - application/x-json-stream
      operationId: "1"
      parameters:
      - description: 文章
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/model.Article'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Result'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Result'
      summary: 提交新的文章内容
      tags:
      - 文章
  /article/{id}:
    delete:
      consumes:
      - application/x-json-stream
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Result'
      summary: 通过id删除指定文章类型
    get:
      consumes:
      - application/x-json-stream
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Result'
      summary: 通过文章 id 获取单个文章内容
  /articles:
    get:
      consumes:
      - application/x-json-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Result'
      summary: 获取所有的文章
swagger: "2.0"